overwrite key symbols via config file
